Fix for JBTM-203
The ReaperThread thread now employs a ReaperWorkerThread to execute
cancel() calls so that a badly behaving TX cannot wedge the reaper. If
the ReaperWorkerThread gets wedged the ReaperThread interrupts it and,
in the absence of any response, marks it as a zombie (i.e. if it ever
unwedges it *has* to exit) and creates a new ReaperWorkerThread.

Failure of a cancel causes the TX to be marked as ROLLBACK_ONLY and an
error to be logged. Failure to mark as ROLLBACK_ONLY causes an error
to be logged. If the (non-exited) zombie count exceeds a threshold an
error is logged.

Defaults for the delays between cancel/interrupt and interrupt/mark as
zombie and for the zombie count threshold can be set via properties.