Enabled use of https for all XTS messages when using the WS-TX 1.1 protocols -- fixes JBTM-347.
This requires enabling of the secure HTTP port in any JBoss AS which
is running either an XTS client, an XST participannt (web service) or
the XTS coordinator. It also requires setting up a keystore and
truststore to allow secure communications using SSL.

Secure communication is switched on by specifying an XTS coordinator
URL with an https scheme. The usual XTS 1.1 coordinator endpoint
address is

  http://<bindhost>:8080/ws-c11/ActivationService

where <bindhost> is the bind address supplied as a parameter on the
JBoss AS run.sh/run.bat command line using the -b flag (<bindhost>
defaults to localhost if this parameter is not supplied).

If the client is run in an AS which specifies the coordinator URL as

  https://<bindhost>:8443/ws-c11/ActivationService

then the initial client<-->coordinator exchange at transaction begin
and all subsequent XTS message exchanges between client<-->coordinator
and participant (web service)<-->coordinator will employ https.

The coordinator URL can be reset in a variety of ways and this fix has
changed the behaviour slightly. One option is to set system
properties, either for the whole URL

  org.jboss.jbossts.xts11.coordinator.url

  if unset then the components properties are tested for and used to
  constrct a URL

or, in case that is unset, by specifying individual components of the
URL

  org.jboss.jbossts.xts11.coordinator.scheme

  value should be http/https and defaults to http


  org.jboss.jbossts.xts11.coordinator.host

  defaults to <bindhost>


  org.jboss.jbossts.xts11.coordinator.port

  defaults to jbossweb http port (normally 8080) if scheme is http and
  jbossweb secure http port (normally 8443) if scheme is https


  org.jboss.jbossts.xts11.coordinator.path

  defaults to ws-c11/ActivationService

If none of these system properties is set then the URL is obtained
from the value of property org.jboss.jbossts.xts11.coordinatorURL
which is specified in property file wstx11.xml which can be found in
config.jar in the deployed jbossxts.sar. Edits to this file take
effect the next time the XTS service is deployed.