Fixes for JBTM-611
reaper check() method now checks for and pulls in pending TXs each
time round the main loop in order to ensure it does not process any
entries out of order. it also locks the queue when it dequeues the
first entry, ensuring that it shares a consistent view of what is
first with any inserting threads

reaper now includes pending TXs when computing transactions count and
timeouts count.

reaper inserts now check the timeouts and pending timeouts maps before
adding an element to detect duplicate adds. This is a guard against
badly implemented hash and equals methods since our code will never
add the same instance twice. Note that a deferred add (to the pending
queue) may succeed but later fail when the TX is added to the sorted
list. This is a guard against badly implemented compareTo
methods. Both of these could be dropped if the performance hit is too
large. We don't really need to care about users who break the reaper
by failing to implement equality or trichotomy.

reaper test script changed to relocate the initial rendezvous point at
the top of the check loop.

ReaperTestCase2 made cosmetic changes so that indicies in symbolic
names for uids in text match indices for actual uids allocated in the
test.