partial changes towards completion of JBTM-741 and JBTM-678
modified WSAS so activities have a service type and only activate high
level services of that type thereby removing some redundant code and
paving the way for simplifying property based config and introducing
bean based config.

modified WSCF and WSTX so that configuration is read from a single
property file rather than multiple xml config files thereby allowing
some of the config code to be deleted and other bits to be simplified.

removed the XML protocol specification documents associated with the
HLS and Context/ContextFactory implementations configured in the
properties file because this indirection did not actually avoid
hard-wiring of the configuration. replaced this XML configuration with
annotations attached to the protocol implementation classes and
modified initalisation code to process these annotations duing
startup. This removes the need for a lot of the remaining
initialisation to be performed.